refactor: remove catalog v1 service
