+++ proto/meta.proto
@@ -279,17 +279,11 @@ message SubscribeResponse {
   uint64 version = 3;
   oneof info {
     common.WorkerNode node = 4;
-    // will be deprecated and replaced by database_v2
-    Database database = 5;
-    // will be deprecated and replaced by schema_v2
-    Schema schema = 6;
-    // will be deprecated and replaced by table_v2
-    Table table = 7;
     catalog.Database database_v2 = 8;
     catalog.Schema schema_v2 = 9;
     catalog.Table table_v2 = 10;
     catalog.Source source = 11;
-    MetaSnapshot fe_snapshot = 12;
+    MetaSnapshot snapshot = 12;
     hummock.HummockSnapshot hummock_snapshot = 13;
   }
 }
