+++ src/meta/src/manager/mod.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-mod catalog;
 mod catalog_v2;
 mod env;
 mod hash_mapping;
@@ -20,7 +19,6 @@ mod id;
 mod notification;
 mod stream_clients;
 
-pub use catalog::*;
 pub use catalog_v2::*;
 pub use env::*;
 pub use hash_mapping::*;
