+++ src/meta/src/rpc/mod.rs
@@ -17,7 +17,6 @@ pub mod metrics;
 pub mod server;
 mod service;
 
-pub use service::catalog_service::CatalogServiceImpl;
 pub use service::cluster_service::ClusterServiceImpl;
 pub use service::ddl_service::DdlServiceImpl;
 pub use service::epoch_service::EpochServiceImpl;
