+++ src/meta/src/rpc/service/mod.rs
@@ -12,7 +12,6 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-pub mod catalog_service;
 pub mod cluster_service;
 pub mod ddl_service;
 pub mod epoch_service;
